tvplanit: Make the results of all calls to GetNextID an Int64 (for consistency with AddResource etc.).
git-svn-id: https://svn.code.sf.net/p/lazarus-ccr/svn@4686 8e941d3f-bd1b-0410-a28a-d453659cc2b4

+++ components/tvplanit/source/vpadvds.pas
@@ -71,7 +71,7 @@ type
     constructor Create(AOwner: TComponent); override;
     destructor Destroy; override;
 
-    function GetNextID(TableName: string): integer; override;
+    function GetNextID(TableName: string): int64; override;
     procedure Load; override;
     procedure CreateTable(TableName: string);
     procedure CreateIndexDefs(const TableName : string;
@@ -220,7 +220,7 @@ begin
 end;
 {=====}
 
-function TVpAdvDataStore.GetNextID(TableName: string): Integer;
+function TVpAdvDataStore.GetNextID(TableName: string): Int64;
 var
   Query: TAdsQuery;
   GotIt: Boolean;
